new battle embed design
