+++ db/mod.ts
@@ -60,7 +60,19 @@ import {
 
 import { createVoteRef, resolveVoteRef } from './voteRef.ts';
 
-export const kv = await Deno.openKv();
+import { load as Dotenv } from '$std/dotenv/mod.ts';
+
+// load .env file
+await Dotenv({
+  export: true,
+  defaultsPath: '.env.example',
+  allowEmptyValues: true,
+  examplePath: null,
+});
+
+export const kv = await Deno.openKv(
+  'https://api.deno.com/databases/c0e82dfc-caeb-4059-877b-3e9134cf6e52/connect',
+);
 
 async function getValue<T>(key: Deno.KvKey): Promise<T | undefined> {
   const res = await kv.get<T>(key);
